refactor: CustomUserDetails 패키지 수장
